Handle float threshold

+++ src/main.ts
@@ -17,7 +17,7 @@ async function run(): Promise<void> {
       return
     }
 
-    const failedThreshold: number = Number.parseInt(core.getInput('failedThreshold'), 10)
+    const failedThreshold: number = Number.parseFloat(core.getInput('failedThreshold'), 10)
     core.debug(`failedThreshold ${failedThreshold}`)
 
     const resultPath: string = core.getInput('resultPath')
